+++ src/aras.jl
@@ -22,13 +22,14 @@ end
 
 
 """
-    aras(decisionMat, weights, fs)
+    aras(decisionMat, weights, fs; normalization)
 Apply ARAS (Additive Ratio ASsessment) method for a given matrix, weights and, type of criteria.
 
 # Arguments:
  - `decisionMat::Matrix`: n × m matrix of objective values for n alternatives and m criteria 
  - `weights::Array{Float64, 1}`: m-vector of weights that sum up to 1.0. If the sum of weights is not 1.0, it is automatically normalized.
  - `fs::Array{<:Function,1}`: m-vector of type of criteria. The benefit criteria shown with "maximum", and the cost criteria shown with "minimum".
+ - `normalization{<:Function}`: Optional normalization function.
 
  # Description 
 aras() applies the ARAS method to rank n alternatives subject to m criteria and criteria type vector.
@@ -75,7 +76,8 @@ Yıldırım, B. F. (2015). "Çok Kriterli Karar Verme Problemlerinde ARAS Yönte
 function aras(
     decisionMat::Matrix,
     weights::Array{Float64,1},
-    fs::Array{F,1}; normalization::G = Normalizations.dividebycolumnsumnormalization 
+    fs::Array{F,1}; 
+    normalization::G = Normalizations.dividebycolumnsumnormalization 
 )::ARASResult where {F<:Function, G<: Function}
 
     # mat = convert(Matrix, decisionMat)