+++ src/cocoso.jl
@@ -26,7 +26,7 @@ end
 
 
 """
-        cocoso(decisionMat, weights, fns; lambda)
+        cocoso(decisionMat, weights, fns; lambda, normalization)
 
 Apply CoCoSo (Combined Compromise Solution) method for a given matrix and weights.
 
@@ -35,6 +35,7 @@ Apply CoCoSo (Combined Compromise Solution) method for a given matrix and weight
  - `weights::Array{Float64, 1}`: m-vector of weights that sum up to 1.0. If the sum of weights is not 1.0, it is automatically normalized.
  - `fns::Array{<:Function, 1}`: m-vector of functions to be applied on the columns.
  - `lambda::Float64`: joint criterion. 0<=lambda<=1, default=0.5.
+ - `normalization{<:Function}`: Optional normalization function.
 
 # Description 
 cocoso() applies the CoCoSo method to rank n alterntives subject to m criteria which are supposed to be