+++ src/codas.jl
@@ -28,7 +28,7 @@ CodasMethod()::CodasMethod = CodasMethod(0.02)
 
 
 """
-    codas(decisionMat, weights, fs)
+    codas(decisionMat, weights, fs; normalization)
 Apply CODAS (COmbinative Distance-based ASsessment) method for a given matrix, weights and, type of criteria.
 
 # Arguments:
@@ -36,6 +36,7 @@ Apply CODAS (COmbinative Distance-based ASsessment) method for a given matrix, w
  - `weights::Array{Float64, 1}`: m-vector of weights that sum up to 1.0. If the sum of weights is not 1.0, it is automatically normalized.
  - `fs::Array{<:Function,1}`: m-vector of type of criteria. The benefit criteria shown with "maximum", and the cost criteria shown with "minimum".
  - `tau::Float64`: tau parameter for the algorithm. The default is 0.02.
+ - `normalization{<:Function}`: Optional normalization function.
 
  # Description 
 codas() applies the CODAS method to rank n alternatives subject to m criteria and criteria type vector.