+++ src/critic.jl
@@ -20,13 +20,14 @@ end
 
 
 """
-        critic(decisionMat, fns)
+        critic(decisionMat, fns; normalization)
 
 Apply CRITIC (Combined Compromise Solution) method for a given matrix and criteria types.
 
 # Arguments:
  - `decisionMat::Matrix`: n × m matrix of objective values for n alternatives and m criteria 
  - `fns::Array{<:Function, 1}`: m-vector of functions to be applied on the columns.
+ - `normalization{<:Function}`: Optional normalization function.
 
 # Description 
 critic() applies the CRITIC method to calculate weights using a decision matrix with