+++ src/mabac.jl
@@ -26,7 +26,7 @@ struct MabacMethod <: MCDMMethod end
 
 
 """
-        mabac(decisionMat, weights, fns)
+        mabac(decisionMat, weights, fns; normalization)
 
 Apply MABAC (Multi-Attributive Border Approximation area Comparison) for a given matrix and weights.
 
@@ -34,6 +34,7 @@ Apply MABAC (Multi-Attributive Border Approximation area Comparison) for a given
  - `decisionMat::Matrix`: n × m matrix of objective values for n alternatives and m criteria 
  - `weights::Array{Float64, 1}`: m-vector of weights that sum up to 1.0. If the sum of weights is not 1.0, it is automatically normalized.
  - `fns::Array{<:Function, 1}`: m-vector of functions to be applied on the columns. 
+ - `normalization{<:Function}`: Optional normalization function.
 
 # Description 
 mabac() applies the MABAC method to rank n alternatives subject to m criteria which are supposed to be