+++ src/mairca.jl
@@ -22,7 +22,7 @@ struct MaircaMethod <: MCDMMethod end
 
 
 """
-        mairca(decisionMat, weights, fns)
+        mairca(decisionMat, weights, fns; normalization)
 
 Apply MAIRCA (Multi Attributive Ideal-Real Comparative Analysis) for a given matrix and weights.
 
@@ -30,6 +30,7 @@ Apply MAIRCA (Multi Attributive Ideal-Real Comparative Analysis) for a given mat
  - `decisionMat::Matrix`: n × m matrix of objective values for n alternatives and m criteria 
  - `weights::Array{Float64, 1}`: m-vector of weights that sum up to 1.0. If the sum of weights is not 1.0, it is automatically normalized.
  - `fns::Array{<:Function, 1}`: m-vector of functions to be applied on the columns. 
+ - `normalization{<:Function}`: Optional normalization function.
 
 # Description 
 mairca() applies the MAIRCA method to rank n alternatives subject to m criteria which are supposed to be