+++ src/psi.jl
@@ -23,13 +23,14 @@ struct PSIMethod <: MCDMMethod end
 
 
 """
-        psi(decisionMat, fns)
+        psi(decisionMat, fns; normalization)
 
 Apply PSI (Preference Selection Index) method for a given matrix and directions of optimizations.
 
 # Arguments:
  - `decisionMat::Matrix`: n × m matrix of objective values for n alterntives and m criteria 
  - `fns::Array{<:Function, 1}`: m-vector of functions to be applied on the columns.
+ - `normalization{<:Function}`: Optional normalization function.
  
 # Description 
 psi() applies the PSI method to rank n alterntives subject to m criteria which are supposed to be