+++ src/saw.jl
@@ -24,7 +24,7 @@ end
 
 
 """
-        saw(decisionMat, weights, fns)
+        saw(decisionMat, weights, fns; normalization)
 
 Apply SAW (Simple Additive Weighting) method for a given matrix and weights.
 This method also known as WSM (Weighted Sum Model)
@@ -33,6 +33,7 @@ This method also known as WSM (Weighted Sum Model)
  - `decisionMat::Matrix`: n × m matrix of objective values for n candidate (or strategy) and m criteria 
  - `weights::Array{Float64, 1}`: m-vector of weights that sum up to 1.0. If the sum of weights is not 1.0, it is automatically normalized.
  - `fns::Array{<:Function, 1}`: m-vector of functions to be applied on the columns. 
+ - `normalization{<:Function}`: Optional normalization function.
 
 # Description 
 saw() applies the SAW method to rank n strategies subject to m criteria which are supposed to be